♻️ Refactor: 사이드 페이지 컴포넌트로 파일 이동
