--- src/pages/Dashboard/components/CardDetail/CommentSection/mockData.js
@@ -1,226 +1,0 @@
-export const mockData = {
-  cursorId: 1,
-  comments: [
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-    {
-      id: 2,
-      content: "댓글을 써보았어요",
-      createdAt: "2024-08-08T09:01:19.170Z",
-      updatedAt: "2024-08-08T09:01:19.170Z",
-      cardId: 1,
-      author: {
-        profileImageUrl:
-          "https://sprint-fe-project.s3.ap-northeast-2.amazonaws.com/taskify/profile_image/4-14_3071_1714403802399.jpeg",
-        nickname: "메타몽",
-        id: 13,
-      },
-    },
-  ],
-};
