+++ src/pages/Dashboard/components/SidePage/CardDetail/CardDetailContent/CardDetailContent.tsx
