+++ src/pages/Dashboard/components/SidePage/CardDetail/CardDetailContent/CardDetailContentStyled.ts
